server/requirements.txt
openenv-core>=0.4.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
numpy>=1.26.0
openai>=1.0.0
